move `contains_nan` to `chainer.backends`
