+++ chainer/backends/__init__.py
@@ -5,6 +5,23 @@ from chainer.backends import cuda
 from chainer.backends import intel64
 
 
+def contains_nan(x):
+    """Returns if x has a NaN value
+
+    Args:
+        x (numpy.ndarray or cupy.ndarray): Array to be checked.
+
+    Returns:
+        bool
+
+    """
+    if x.dtype.kind in ('f', 'c'):
+        with cuda.get_device_from_array(x)
+            return cuda.get_array_module(x).isnan(x).any()
+    else:
+        return False
+
+
 def copyto(dst, src):
     """Copies the elements of an ndarray to those of another one.
