+++ chainer/utils/__init__.py
@@ -6,7 +6,6 @@ import tempfile
 import numpy
 
 # import classes and functions
-from chainer.utils.array import contains_nan  # NOQA
 from chainer.utils.array import sum_to  # NOQA
 from chainer.utils.conv import get_conv_outsize  # NOQA
 from chainer.utils.conv import get_deconv_outsize  # NOQA
