+++ chainer/utils/array.py
@@ -39,23 +39,6 @@ def empty_like(x):
         return numpy.empty_like(x)
 
 
-def contains_nan(x):
-    """Returns if x has a NaN value
-
-    Args:
-        x (numpy.ndarray or cupy.ndarray): Array to be checked.
-
-    Returns:
-        bool
-
-    """
-    if x.dtype.kind in ('f', 'c'):
-        with cuda.get_device_from_array(x)
-            return cuda.get_array_module(x).isnan(x).any()
-    else:
-        return False
-
-
 def sum_to(x, shape):
     if x.shape == shape:
         return x
